+++ src/battle/battle_impact_emitters.c
@@ -6,7 +6,7 @@ enum BattleImpactEmitterOffset {
 };
 
 BattleAITask *BattleImpactEmitter_Start(
-    u16 object_id, int mode, s16 target_scale_q4, u16 lifetime_multiplier,
+    u16 object_id, int mode, int target_scale_q4, int lifetime_multiplier,
     int spawn_interval, int size_multiplier, int target_size,
     int initial_z_offset, int z_velocity, int reserved) {
     BattleImpactTrailEmitterTask *task =
